Fjernet gammel kode
