+++ apps/testnorge-bisys/build.gradle
@@ -71,7 +71,6 @@ dependencies {
 
     implementation 'org.springdoc:springdoc-openapi-ui:1.6.1'
     implementation 'net.logstash.logback:logstash-logback-encoder:7.0.1'
-    implementation 'no.nav.registre:testnorge-consumers:1.8.0-RELEASE'
 
     implementation 'com.squareup.okhttp3:okhttp:4.9.3'
