+++ apps/testnorge-bisys/src/main/java/no/nav/registre/bisys/config/AppConfig.java
@@ -1,14 +1,10 @@
 package no.nav.registre.bisys.config;
 
-import no.nav.registre.testnorge.consumers.hodejegeren.HodejegerenConsumer;
 import no.nav.testnav.libs.servletcore.config.ApplicationCoreConfig;
 import no.nav.testnav.libs.standalone.servletsecurity.config.InsecureJwtServerToServerConfiguration;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
+
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.DependsOn;
 import org.springframework.context.annotation.Import;
-import org.springframework.web.client.RestTemplate;
 
 @Configuration
 @Import(value = {
@@ -16,18 +12,4 @@ import org.springframework.web.client.RestTemplate;
         InsecureJwtServerToServerConfiguration.class
 })
 public class AppConfig {
-
-    @Value("${consumers.hodejegeren.url}")
-    private String hodejegerenUrl;
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-
-    @Bean
-    @DependsOn("restTemplate")
-    public HodejegerenConsumer hodejegerenConsumer() {
-        return new HodejegerenConsumer(hodejegerenUrl, restTemplate());
-    }
 }
