--- apps/testnorge-bisys/src/main/java/no/nav/registre/bisys/config/AutoTimingConfiguration.java
@@ -1,17 +1,0 @@
-package no.nav.registre.bisys.config;
-
-import io.micrometer.core.aop.TimedAspect;
-import io.micrometer.core.instrument.MeterRegistry;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@Configuration
-@EnableAspectJAutoProxy
-public class AutoTimingConfiguration {
-
-    @Bean
-    public TimedAspect timedAspect(MeterRegistry registry) {
-        return new TimedAspect(registry);
-    }
-}
