--- apps/testnorge-bisys/src/main/java/no/nav/registre/bisys/exception/SyntetisertBidragsmeldingException.java
@@ -1,9 +1,0 @@
-package no.nav.registre.bisys.exception;
-
-public class SyntetisertBidragsmeldingException extends Exception {
-
-    public SyntetisertBidragsmeldingException(String message) {
-        super(message);
-    }
-
-}
