+++ apps/testnorge-bisys/src/test/java/no/nav/registre/bisys/ApplicationContextTest.java
@@ -1,23 +1,21 @@
 package no.nav.registre.bisys;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.test.context.ActiveProfiles;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureWireMock(port = 0)
-@TestPropertySource(locations = "classpath:application-test.properties")
-public class ApplicationContextTest {
+@SpringBootTest
+@ActiveProfiles("test")
+class ApplicationContextTest {
 
-    /**
-     * Used to test app runtime dependency and properties
-     */
+    @MockBean
+    public JwtDecoder jwtDecoder;
+
     @Test
     @SuppressWarnings("java:S2699")
-    public void load_app_context() {
+    void load_app_context() {
     }
 }
+
