+++ apps/testnorge-bisys/src/test/resources/application-test.yml
@@ -18,10 +18,5 @@ consumers:
     cluster: dummt
     namespace: dummy
     name: synthdata-bisys
-  person-search-service:
-    url: http://localhost:${wiremock.server.port:0}/person-search
-    cluster: dummy
-    namespace: dummy
-    name: person-search-service
 
 AAD_ISSUER_URI: http://localhost:${wiremock.server.port:0}/token
